+++ include/mat.h
@@ -8,7 +8,7 @@ namespace Geometry
   public:
     static Mat identity();
     Mat(){}
-    Mat(IntVector v1,IntVector v2, IntVector v3);
+    Mat(FloatVector v1,FloatVector v2, FloatVector v3);
     Mat(int a00,int a01,int a02,int a10,int a11,int a12,int a20,int a21,int a22);
     const double el(int i, int j) const
     {
@@ -23,7 +23,7 @@ namespace Geometry
     Mat inverse();
     bool operator == (const Mat& m) const;
     Mat operator * (const Mat& m) const;
-    IntVector operator * (const IntVector& v) const;
+    FloatVector operator * (const FloatVector& v) const;
   private:
     double ar[9];
   };
