+++ include/puzzlepartshower.h
@@ -24,14 +24,14 @@ namespace Visualization
 		class Triangle
 		{
 		public:
-			Triangle(Geometry::IntVector vert1, Geometry::IntVector vert2,Geometry::IntVector vert3, Geometry::IntVector ni):v1(vert1),v2(vert2),v3(vert3),n(ni){}
-			Geometry::IntVector v1,v2,v3,n;
+			Triangle(Geometry::FloatVector vert1, Geometry::FloatVector vert2,Geometry::FloatVector vert3, Geometry::FloatVector ni):v1(vert1),v2(vert2),v3(vert3),n(ni){}
+			Geometry::FloatVector v1,v2,v3,n;
 		};
 		class Plane
 		{
 		public:
-			Plane(Geometry::IntVector pi, Geometry::IntVector ni):p(pi),n(ni){}
-			Geometry::IntVector p,n;
+			Plane(Geometry::FloatVector pi, Geometry::FloatVector ni):p(pi),n(ni){}
+			Geometry::FloatVector p,n;
 		};
 		bool couldDraw_(const Geometry::VolPart& v, Triangle tri) const;
 		bool triIntersectsPlane_(Triangle tri, Plane p) const;
